moved CVEProject description grabber to separate file and added internal cache to it
